make lyric caret position state into single class.
